Add IP List Feature
And removed IPViewer on the target side, added fullscreen mode for the stream viewer, and the stream viewer can now use like a typing zone, it only works on fullscreen mode though. Unoptimized? the stream sender and receiver but somehow being a little faster. And finally, the IP List can store a name of the target, it store in BinaryConfig format at %USER_HOME%\controlax.bcfg